requirements.txt
pandas
numpy
scikit-learn
matplotlib
opencv-python
Pillow
pyyaml
